requirements.txt
python-pptx
pywin32
PyPDF2 